DateFormatterMixin and AlertMixin implemented
