+++ app/javascript/components/rules/RuleHistories.vue
@@ -31,8 +31,11 @@
 
 <script>
 import axios from 'axios';
+import DateFormatMixinVue from '../../mixins/DateFormatMixin.vue';
+import AlertMixinVue from '../../mixins/AlertMixin.vue';
 export default {
   name: 'RuleHistories',
+  mixins: [DateFormatMixinVue, AlertMixinVue],
   props: {
     rule: {
       type: Object,
